Time: 120 ms (12.31%), Space: 89.3 MB (39.21%) - LeetHub

+++ 2364-count-number-of-bad-pairs/2364-count-number-of-bad-pairs.cpp
@@ -13,7 +13,7 @@ public:
         }
         long long count=0;
         for(auto pair:m){
-            int x=pair.second;
+            long long x=pair.second;
             count+=(x*(x-1))/2;
         }
